+++ core/src/main/java/com/populaire/projetguerrefroide/economy/good/ResourceProductionType.java
@@ -7,7 +7,7 @@ import com.populaire.projetguerrefroide.economy.population.PopulationType;
 import java.util.List;
 
 public class ResourceProductionType extends ProductionType {
-    public ResourceProductionType(short workforce, PopulationType owner, List<Employee> employees) {
+    public ResourceProductionType(int workforce, PopulationType owner, List<Employee> employees) {
         super(workforce, owner, employees);
     }
